+++ packages/edition-slices/__tests__/tile-ae/shared-tile-ae.base.js
@@ -1,0 +1,24 @@
+import "../mocks-tiles";
+import { editionBreakpoints } from "@times-components/styleguide";
+import { testTile } from "../shared-tile-utils";
+import { TileAE } from "../../src/tiles";
+
+export default () => {
+  describe("tile ae", () => {
+    it("medium", () => {
+      testTile(TileAE, editionBreakpoints.medium);
+    });
+
+    it("wide", () => {
+      testTile(TileAE, editionBreakpoints.wide);
+    });
+
+    it("huge", () => {
+      testTile(TileAE, editionBreakpoints.huge);
+    });
+
+    it("without breakpoint should be like medium", () => {
+      testTile(TileAE);
+    });
+  });
+};