+++ packages/edition-slices/__tests__/web/tile-al.test.js
@@ -1,0 +1,3 @@
+import shared from "../tile-al/shared-tile-al.web";
+
+shared();