+++ packages/edition-slices/__tests__/web/tile-h.test.js
@@ -1,0 +1,3 @@
+import shared from "../tile-h/shared-tile-h.web";
+
+shared();