+++ packages/edition-slices/__tests__/web/tile-m.test.js
@@ -1,0 +1,3 @@
+import shared from "../tile-m/shared-tile-m.web";
+
+shared();