+++ packages/edition-slices/__tests__/web/tile-u.test.js
@@ -1,0 +1,3 @@
+import shared from "../tile-u/shared-tile-u.web";
+
+shared();