+++ packages/edition-slices/__tests__/web/tile-w.test.js
@@ -1,0 +1,3 @@
+import shared from "../tile-w/shared-tile-w.web";
+
+shared();