+++ packages/ssr/src/standalone-renderer/app.js
@@ -92,6 +92,7 @@ server.get("/article/:id", (request, response) => {
       headers,
       {
         ...makeUrls,
+        enableNewskit: request.param("newskit"),
         graphqlApiUrl,
         logger,
         spotAccountId